Refactored some classes and fields for athletics feature. Added layouts for athletics feature
