--- res/layout/news_items.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!-- Created by Jason Liu on 10/04/14.
-     This layout is used in NewsChooser and NewsReader, except the title is hidden in
-     NewsReader. It is either a list of RSS feeds or a list of RSS items, which are
-     populated into the ListView listNews below
--->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/listNews"
-        android:layout_gravity="center_horizontal"/>
-</LinearLayout>
