+++ res/layout/rss_items.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!-- Created by Jason Liu on 10/04/14.
+     This layout is used in NewsChooser, NewsReader, AthleticsChooser, & Headlines.
+     It is either a list of RSS feeds or a list of RSS items, which are
+     populated into the ListView listNews below.
+-->
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ListView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/listRSS"
+        android:layout_gravity="center_horizontal"/>
+
+    <TextView
+        android:id="@android:id/empty"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_gravity="center"
+        android:gravity="center"
+        android:text="Nothing to show"
+        android:textSize="30sp"
+        android:textStyle="bold"/>
+</LinearLayout>
