+++ res/menu/athletics_toggle.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/toggle"
+          android:title="Scores"
+          android:showAsAction="always"
+          />
+</menu>
