+++ res/values/strings.xml
@@ -17,7 +17,7 @@
         <item>Athletics</item>
         <item>Arts Events</item>
     </string-array>
-    <string-array name="rss_feeds">
+    <string-array name="news_feeds">
         <item>http://news.yale.edu/news-rss</item>
         <item>http://news.yale.edu/topics/arts-humanities/rss</item>
         <item>http://news.yale.edu/topics/business-law-society/rss</item>
@@ -25,7 +25,7 @@
         <item>http://news.yale.edu/topics/science-health/rss</item>
         <item>http://news.yale.edu/topics/world-environment/rss</item>
     </string-array>
-    <string-array name="rss_names">
+    <string-array name="news_names">
         <item>All Topics</item>
         <item>Arts and Humanities</item>
         <item>Business, Law, Society</item>
