+++ src/edu/yalestc/yalepublic/RSS/RssReader.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package edu.yalestc.yalepublic.News;
+package edu.yalestc.yalepublic.RSS;
 
 import java.io.IOException;
 import java.io.InputStream;
